# twitter-clone-backend/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:61326","basePath":"/","tags":[{"name":"tweets-controller","description":"Tweets Controller"}],"paths":{"/tweets":{"post":{"tags":["tweets-controller"],"summary":"post","operationId":"postUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"tweetRequest","description":"tweetRequest","required":true,"schema":{"$ref":"#/definitions/TweetRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Tweet"}}}},"delete":{"tags":["tweets-controller"],"summary":"this is a utility method, please ignore, just added for my own convenience so I don't need to create a bastion to run sql scripts to truncate the tweets table","operationId":"deleteAllUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/tweets/profile":{"get":{"tags":["tweets-controller"],"summary":"profile","operationId":"profileUsingGET","produces":["*/*"],"parameters":[{"name":"offset","in":"query","required":false,"type":"integer","format":"int64"},{"name":"paged","in":"query","required":false,"type":"boolean"},{"name":"pageNumber","in":"query","required":false,"type":"integer","format":"int32"},{"name":"pageSize","in":"query","required":false,"type":"integer","format":"int32"},{"name":"sort.sorted","in":"query","required":false,"type":"boolean"},{"name":"sort.unsorted","in":"query","required":false,"type":"boolean"},{"name":"unpaged","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Page«Tweet»"}}}}},"/tweets/timeline":{"get":{"tags":["tweets-controller"],"summary":"timeline","operationId":"timelineUsingGET","produces":["*/*"],"parameters":[{"name":"offset","in":"query","required":false,"type":"integer","format":"int64"},{"name":"paged","in":"query","required":false,"type":"boolean"},{"name":"pageNumber","in":"query","required":false,"type":"integer","format":"int32"},{"name":"pageSize","in":"query","required":false,"type":"integer","format":"int32"},{"name":"sort.sorted","in":"query","required":false,"type":"boolean"},{"name":"sort.unsorted","in":"query","required":false,"type":"boolean"},{"name":"unpaged","in":"query","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Page«Tweet»"}}}}}},"securityDefinitions":{"JWT Bearer token":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"Pageable":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"sort":{"$ref":"#/definitions/Sort"},"unpaged":{"type":"boolean"}},"title":"Pageable"},"Page«Tweet»":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/definitions/Tweet"}},"empty":{"type":"boolean"},"first":{"type":"boolean"},"last":{"type":"boolean"},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/definitions/Pageable"},"size":{"type":"integer","format":"int32"},"sort":{"$ref":"#/definitions/Sort"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}},"title":"Page«Tweet»"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"},"Sort":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}},"title":"Sort"},"Tweet":{"type":"object","properties":{"author":{"$ref":"#/definitions/User"},"content":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"id":{"type":"string","format":"uuid"}},"title":"Tweet"},"TweetRequest":{"type":"object","properties":{"content":{"type":"string","minLength":0,"maxLength":160}},"title":"TweetRequest"},"User":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}